libstdc++: Check for multiple modifiers in chrono format string [PR110708]
Tested x86_64-linux. Pushed to trunk.

This should be backported to gcc-13 too, but it can wait until 13.3
(it's just an accepts-invalid and unlikely to affect anybody in
practice).

-- >8 --

The logic for handling modified chrono specs like %Ey was just
restarting the loop after each modifier, and not checking whether we'd
already seen a modifier.

libstdc++-v3/ChangeLog:

	PR libstdc++/110708
	* include/bits/chrono_io.h (__formatter_chrono::_M_parse): Only
	allow a single modifier.
	* testsuite/std/time/format.cc: Check multiple modifiers.