c++: Improve constexpr error for dangling local variables [PR110619]
Currently, when typeck discovers that a return statement will refer to a
local variable it rewrites to return a null pointer. This causes the
error messages for using the return value in a constant expression to be
unhelpful, especially for reference return values.

This patch removes this "optimisation". Relying on this raises a warning
by default and causes UB anyway, so there should be no issue in doing
so. We also suppress additional warnings from later passes that detect
this as a dangling pointer, since we've already indicated this anyway.

	PR c++/110619

gcc/cp/ChangeLog:

	* semantics.cc (finish_return_stmt): Suppress dangling pointer
	reporting on return statement if already reported.
	* typeck.cc (check_return_expr): Don't set return expression to
	zero for dangling addresses.

gcc/testsuite/ChangeLog:

	* g++.dg/cpp1y/constexpr-lifetime5.C: Test reported message is
	correct.
	* g++.dg/cpp1y/constexpr-lifetime6.C: Likewise.
	* g++.dg/cpp1y/constexpr-110619.C: New test.
	* g++.dg/warn/Wreturn-local-addr-6.C: Remove check for return
	value optimisation.

Signed-off-by: Nathaniel Shead <nathanieloshead@gmail.com>