LoongArch: Allow using --with-arch=native if host CPU is LoongArch
If the host triple and the target triple are different but the host is
LoongArch, in some cases --with-arch=native can be useful.  For example,
if we are bootstrapping a loongarch64-linux-musl toolchain on a
Glibc-based system and we don't intend to use the toolchain on other
machines, we can use

    ../gcc/configure --{build,host}=loongarch64-linux-gnu \
                     --target=loongarch64-linux-musl --with-arch=native

Relax the check in config.gcc to allow such configurations.

gcc/ChangeLog:

	* config.gcc [target=loongarch*-*-*, with_arch=native]: Allow
	building cross compiler if the host CPU is LoongArch.