Fix length computation bug in bfin port
The tester seemed to occasionally ping-pong a compilation failure on the
builtin-bitops-1.c test.  I long suspected it was something like length
computations.

I finally got a few minutes to dig into it, and sure enough the blackfin
port was claiming the "ones" operation was 2 bytes when it is in fact 4
bytes.

This fixes the compilation failure for the builtin-bitops-1.c test.
Sadly, it doesn't fix any of the other failures on the bfin port.

Committed to the trunk.

Jeff
commit bb095e8a343db043a0cd0b0da9b2ab1186d1a1ed
Author: Jeff Law <jlaw@ventanamicro.com>
Date:   Sat Jul 22 09:47:21 2023 -0600

    [committed] Fix length computation bug in bfin port

    The tester seemed to occasionally ping-pong a compilation failure on the
    builtin-bitops-1.c test.  I long suspected it was something like length
    computations.

    I finally got a few minutes to dig into it, and sure enough the blackfin
    port was claiming the "ones" operation was 2 bytes when it is in fact 4 bytes.

    This fixes the compilation failure for the builtin-bitops-1.c test.   Sadly,
    it doesn't fix any of the other failures on the bfin port.

    Committed to the trunk.

    gcc/
            * config/bfin/bfin.md (ones): Fix length computation.