libstdc++: Reuse double overload of __convert_to_v if possible
This seems like a worthwhile change. In theory we could make it even
smaller by using a symbol alias, but I'm not sure that's supported on
the targets where this will actually help.

Any objections?

-- >8 --

For targets where double and long double have the same representation we
can reuse the same __convert_to_v code for both types. This will
slightly reduce the size of the compiled code in the library.

libstdc++-v3/ChangeLog:

	* config/locale/generic/c_locale.cc (__convert_to_v): Reuse
	double overload for long double if possible.